requirements/base.txt
asgiref==3.8.1
Django==5.1.5
django-bootstrap5==24.3
django-debug-toolbar==5.0.1
django-taggit==6.1.0
django-tinymce==4.1.0
djangorestframework==3.15.2
pillow==11.1.0
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
